WIP: Constrain registers within a range
Certain hardware features will limit the number of registers available
to instructions. E.g., Intel APX expands the number of GPRs to 32, but
certain instructions are not APX-extended. In these cases, we want to
limit the register set to the range `0-15`. This can happen similarly
with vector registers: if we were to use all 32 XMMs allowed by EVEX but
tried to mix in SSE instructions, these SSE instructions would need to
be limited to the range `0-15`.

This change adds a new `OperandConstraint` that restricts which
registers are available for allocation. Currently, the range is limited
to `0..n` where `n` is a power of two. This could be expanded in the
future with some refactoring of the `Operand` bit encodings.

This is a WIP: it is unclear if updating the allocation logic is enough.
There are various expressions like `if let
OperandConstraint::FixedReg(preg) = ...` that may need to be audited.