Refactor `RegTraversalIter::new` parameter order
This does not change behavior, but simply passes through the arguments
constructing a `RegTraversalIter` in the order they will be checked
during iteration: first it emits the fixed register, then the hinted
register, then the offset-based lookup.

Another possible change here would be to pass the hinted register as an
`Option<PReg>` with `None` to indicate no hint. A quick audit of the
code supports this but I avoided this out of an abundance of caution.