Generate arbitrary functions with limit constraints
This gives the fuzzer the ability to generate functions with
`OperandConstraint::Limit`. This is a tricky proposition, since now
there is yet another way that generated functions can fail to allocate
beyond the interplay of too many live values, fixed registers, and
clobbers. The current generation relies on some magic maximums (i.e., 10
operands, 10 clobbers); with 32 registers available, we can always
allocate that. But adding limits to this equation reduces the number of
available registers.